+++ tests/test_app/TestApp/config/banned_ip.php
@@ -1,6 +1,0 @@
-<?php
-//Array of banned IP addresses.
-//You can use the asterisk (*) as a wildcard.
-//With an empty value, access is granted to any ip addresses (no limitation).
-return ['Banned' => [
-]];