+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.26 branch
 ### 2.26.5
+* it uses the `cakephp-stop-spam` package to detect spammers;
+* `AppController::isBanned()` method renamed as `isSpammer()`;
 * updated for `me-tools` 2.18.7.
 
 ### 2.26.4